+++ src/BaseClient.ts
@@ -60,8 +60,8 @@ export function normalizeClientOptions<T extends BaseClientOptions = BaseClientO
         {
             "X-Fern-Language": "JavaScript",
             "X-Fern-SDK-Name": "phenoml",
-            "X-Fern-SDK-Version": "17.5.0",
-            "User-Agent": "phenoml/17.5.0",
+            "X-Fern-SDK-Version": "17.6.0",
+            "User-Agent": "phenoml/17.6.0",
             "X-Fern-Runtime": core.RUNTIME.type,
             "X-Fern-Runtime-Version": core.RUNTIME.version,
         },